ver 1.3

+++ protected/controllers/TaskController.php
@@ -1,24 +1,36 @@
 <?php
 
+/**
+ * Created by PhpStorm.
+ * User: Ghazi
+ * Date: 24/02/2015
+ * Time: 16:00 PM
+ */
 class TaskController extends Controller
 {
-    /* edited by ebrahim ver 1.1 */
+    //1.0.2 added filter and order but not complete
     public function actionIndex()
     {
         if (!Yii::app()->user->isGuest) {
-            RoleHelper::checkAccessControl('task', 'view', null, true, false);
-            $model = TaskHelper::getModel();
+            RoleHelper::checkAccessControl('task', 'view', $this, true, false);
+
+            if (isset($_GET['status'])) {
+                $model = TaskHelper::getModel($_GET['status']);
+            } else {
+                $model = TaskHelper::getModel();
+            }
             $this->render('index', array('model' => $model));
             LogHelper::proccess(LogHelper::VIEW, LogHelper::TASK, "مشاهده ی تسک ها");
+
         } else {
             $this->redirect(array('site/login'));
         }
     }
 
-    /* edited by ebrahim ver 1.1 */
     public function actionNew()
     {
-        RoleHelper::checkAccessControl('task', 'insert');
+        RoleHelper::checkAccessControl('task', 'insert', $this);
+
         $model = new Task;
 
         if (isset($_POST['ajax']) && $_POST['ajax'] === 'task-new-form') {
@@ -49,7 +61,33 @@ class TaskController extends Controller
         }
         $this->renderPartial('new', array('model' => $model, false, true));
     }
-    /* edited by ebrahim ver 1.1 */
+
+    public function actionNew_group()
+    {
+        RoleHelper::checkAccessControl('task', 'insert', $this);
+        $model = new GroupTask;
+
+        if (isset($_POST['ajax']) && $_POST['ajax'] === 'group-new-form') {
+            echo CActiveForm::validate($model);
+            Yii::app()->end();
+        }
+
+        if (isset($_POST['GroupTask'])) {
+            $model->attributes = $_POST['GroupTask'];
+            $model->time = time();
+            $model->user_id = Yii::app()->user->id;
+
+            if ($model->validate()) {
+                if ($model->save(false)) {
+                    LogHelper::proccess(LogHelper::INSERT_GROUP, LogHelper::TASK, "گروه جدید اضافه شد");
+                }
+                $this->renderPartial('/site/success', null, false, true);
+                Yii::app()->end();
+            }
+        }
+        $this->renderPartial('new_group', array('model' => $model, false, true));
+    }
+
     public function actionEdit()
     {
         RoleHelper::checkAccessControl('task', 'edit', $this);
@@ -83,7 +121,7 @@ class TaskController extends Controller
             $this->renderPartial('edit', array('model' => $model, false, true));
         }
     }
-    /* edited by ebrahim ver 1.1 */
+
     public function actionInfo()
     {
         RoleHelper::checkAccessControl('task', 'view', $this);
@@ -92,40 +130,53 @@ class TaskController extends Controller
         }
         $this->renderPartial('info', array('model' => $model, false, true));
     }
-    /* edited by ebrahim ver 1.1 */
+
     public function actionDel()
     {
 
-        RoleHelper::checkAccessControl('task', 'delete', $this);
-        if (isset($_POST['id'])) {
-            $id = $_POST['id'];
-            $task = Task::model()->findByPk($id);
-            if ($task->delete()) {
-                LogHelper::proccess(LogHelper::DELETE, LogHelper::TASK, "تسک حذف شد");
+        if (RoleHelper::checkAccessControl('task', 'delete')) {
+            if (isset($_POST['id'])) {
+                $id = $_POST['id'];
+                $task = Task::model()->findByPk($id);
+                if ($task->delete()) {
+                    LogHelper::proccess(LogHelper::DELETE, LogHelper::TASK, "تسک حذف شد");
+                }
+                $this->renderPartial('/site/success', null, false, true);
             }
-            $this->renderPartial('/site/success', null, false, true);
+        } else {
+            $this->renderPartial('/site/noaccess', null, false, true);
         }
     }
 
-    public function filters()
+    //1.0.2 this returns all tasks of a spacial project
+    public function actionProjecttask()
     {
-        return array(
-            'accessControl',
-        );
+        $message = "";
+        if (RoleHelper::checkAccessControl('task', 'view')) {
+            if (isset($_GET['id'])) {
+
+                $model = Task::model()->findAll("id_project=:id", array(':id' => $_GET['id']));
+            }
+            $this->render('Projecttask', array('model' => $model, 'title' => $_GET['title'], 'message' => $message));
+        } else {
+            $message = "شما اجازه ی مشاهده ی تسک ها را ندارید";
+            $this->render('index', array('message' => $message));
+        }
     }
 
-    public function accessRules()
+    //1.0.2 added print
+    public function actionprint()
     {
-        return array(
-            array('allow',
-                'users' => array('@'),
-            ),
-            array('deny',
-                'users' => array('*'),
-            ),
-        );
-    }
 
+        if (isset($_GET['id'])) {
+            $model = Task::model()->findByPk($_GET['id']);
+            if ($model != null) {
+                $this->render('print', array('model' => $model));
+            } else {
+                $this->render('print', array('model' => array()));
+            }
+        }
+    }
     // Uncomment the following methods and override them if needed
     /*
     public function filters()
